requirements.txt
# MOSAR GraphRAG Requirements Management System
# Python Dependencies

# Core
python-dotenv==1.0.0

# Neo4j
neo4j==5.14.1

# OpenAI
openai==1.12.0

# Embeddings & NLP
sentence-transformers==2.3.1
scikit-learn==1.4.0
numpy==1.26.3

# Data Processing
pandas==2.2.0
openpyxl==3.1.2

# Utilities
tqdm==4.66.1
